requirements.txt
hidapi==0.15.0
pynput==1.8.1
pyobjc-core==12.1
pyobjc-framework-ApplicationServices==12.1
pyobjc-framework-Cocoa==12.1
pyobjc-framework-CoreText==12.1
pyobjc-framework-Quartz==12.1
